# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Minha API de Reservas",
        "description": "Documentação da API v1",
        "contact": {
            "name": "Suporte API",
            "email": "suporte@seudominio.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registrar novo usuário",
                "operationId": "38e3071463848ba373dfeae5623cdfa4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "User Name"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validação falhou"
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login de usuário via API",
                "operationId": "b9b760991d94b179960cd7ddc93eedf2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
        "/api/v1/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Retorna o usuário autenticado",
                "operationId": "c24e88aea4324667c5f4e38549a5298f",
                "responses": {
                    "200": {
                        "description": "Usuário retornado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout do usuário (token inválido)",
                "operationId": "b66dd8a7864f561822c01bc8339363fb",
                "responses": {
                    "200": {
                        "description": "Logout realizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logout realizado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/reservations": {
            "post": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Cria uma nova reserva",
                "operationId": "0a898a039908af554f99719b597122de",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "room_id",
                                    "reservation_date",
                                    "reservation_time"
                                ],
                                "properties": {
                                    "room_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "reservation_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-12"
                                    },
                                    "reservation_time": {
                                        "type": "string",
                                        "example": "14:30"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pendente",
                                            "concluido",
                                            "cancelado"
                                        ],
                                        "example": "pendente"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reserva criada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reserva criada com sucesso!"
                                        },
                                        "reservation": {
                                            "$ref": "#/components/schemas/Reservation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/rooms": {
            "get": {
                "tags": [
                    "Salas"
                ],
                "summary": "Lista todas as salas",
                "operationId": "8621bf9d381104070fb98293d57bc3c2",
                "responses": {
                    "200": {
                        "description": "Lista de salas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Room"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Reservation": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "room_id": {
                        "type": "integer"
                    },
                    "reservation_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "reservation_time": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pendente",
                            "concluido",
                            "cancelado"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Room": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "room_name": {
                        "type": "string"
                    },
                    "location_name": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "capacity": {
                        "type": "integer"
                    },
                    "description": {
                        "type": "string"
                    },
                    "rating": {
                        "type": "integer"
                    },
                    "image": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "role": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Reservas",
            "description": "Reservas"
        },
        {
            "name": "Salas",
            "description": "Salas"
        }
    ]
}